+++ lib/src/models/user_menu_item.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/widgets.dart';
+
+/// A single item in the user flyout menu.
+class UserMenuItem {
+  const UserMenuItem({
+    required this.label,
+    this.icon,
+    required this.onTap,
+  });
+
+  final String label;
+  final IconData? icon;
+  final VoidCallback onTap;
+}